Refactored command parsing and execution into command specific classes
Previously all code regarding parsing and execution of command lived in CommandRunner. That class and testing of it was getting out of hand wiht new commands needing to be added soon. So made a decision to factor out the code for each command into thier own class. This class will be responsible for parsing the command as well as executing the parsed input.

The interface to all comands is homogenous. All of them get an image parameter which may or may not have been initialised. And they get any parsed parameters as hash. they can choose to ignore parts of hash that is being passed ot them.